Change comparison order in make.sh diff
The order two files are compared against each other in when running
the command `make.sh diff` is swapped for a more intuitive
reading. Indeed, by performing `diff <old-file> <new-file>` instead of
the other way around, the output reads as a patch representing what we
happen if we "apply" the newly generated dotfiles to the existing
ones.

Signed-off-by: Antoine Busque <antoine.busque@pillrcompany.com>